Configure rabbitmq with frame_size 0
After an update to a higher rabbitmq-server version, we need to run rabbitmq
with frame_size set to 0. We are limited to use a bunny version older than 0.9
that does not support frames yet, causing a rejected connecions on chef-server.

After updating to a newer version of chef and bunny, this can be removed.
See also https://bugzilla.suse.com/show_bug.cgi?id=910815